+++ foo_ui_columns/playlist_switcher_title_formatting.h
@@ -2,6 +2,6 @@
 
 namespace cui {
 
-pfc::string8 format_playlist_title(unsigned index);
+pfc::string8 format_playlist_title(size_t index);
 
 }
